+++ Sources/MapboxMaps/Style/AttributionDialogManager.swift
@@ -1,3 +1,4 @@
+import UIKit
 internal protocol AttributionDataSource: AnyObject {
     func attributions() -> [Attribution]
 }
@@ -131,11 +132,18 @@ extension AttributionDialogManager: InfoButtonOrnamentDelegate {
         let bundle = Bundle.mapboxMaps
 
         if let attributions = dataSource?.attributions() {
-            for attribution in attributions {
-                let action = UIAlertAction(title: attribution.title, style: .default) { _ in
-                    self.delegate?.attributionDialogManager(self, didTriggerActionFor: attribution)
+
+            // Non actionable single item gets displayed as alert's message
+            if attributions.count == 1, let attribution = attributions.first, attribution.kind == .nonActionable {
+                alert.message = attribution.title
+            } else {
+                for attribution in attributions {
+                    let action = UIAlertAction(title: attribution.title, style: .default) { _ in
+                        self.delegate?.attributionDialogManager(self, didTriggerActionFor: attribution)
+                    }
+                    action.isEnabled = attribution.kind != .nonActionable
+                    alert.addAction(action)
                 }
-                alert.addAction(action)
             }
         }