+++ Tests/MapboxMapsTests/Style/Mocks/MockAttributionDataSource.swift
@@ -1,0 +1,11 @@
+import XCTest
+@testable import MapboxMaps
+import Foundation
+
+final class MockAttributionDataSource: AttributionDataSource {
+
+    let attributionsStub = Stub<Void, [Attribution]>(defaultReturnValue: [])
+    func attributions() -> [Attribution] {
+        attributionsStub.call()
+    }
+}